--- backend/src/websocket/websocket.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { WebsocketService } from './websocket.service';
-
-describe('WebsocketService', () => {
-  let service: WebsocketService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [WebsocketService],
-    }).compile();
-
-    service = module.get<WebsocketService>(WebsocketService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
